backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "google-genai==1.57.0",
    "pinecone==8.0.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "uvicorn[standard]==0.34.0",
]
